Split out various code gen bits to separate files for better readability
