+++ circuit-codegen/src/main/kotlin/com/slack/circuit/codegen/FactoryType.kt
@@ -1,0 +1,6 @@
+package com.slack.circuit.codegen
+
+internal enum class FactoryType {
+  PRESENTER,
+  UI
+}
